FCQ/quickCheck_httpx/requirements.txt
pyyaml==6.0
httpx==0.23.3